% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nba_player_stats.R
\name{scrape_nba_player_stats}
\alias{scrape_nba_player_stats}
\title{Scrapes NBA player stats tables off basketball-reference.com.}
\usage{
scrape_nba_player_stats(name, stats_tb)
}
\arguments{
\item{name}{is a char string that corresponds to the players name.}

\item{stats_tb}{is a char string that corresponds to the statistics table such as #per_game, #totals, #per_36_minutes, and #advanced.}
}
\value{
a data frame of statistics for a specific NBA player.
}
\description{
Scrapes NBA player stats tables off basketball-reference.com.
}
\examples{
scrape_nba_player_stats("Allen Iverson", "#per_game")
}
