% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nba_team_stats.R
\name{scrape_nba_team_stats}
\alias{scrape_nba_team_stats}
\title{Scrapes NBA team stats tables off basketball-reference.com.}
\usage{
scrape_nba_team_stats(slug, year, stats_tb)
}
\arguments{
\item{slug}{is a char string that corresponds to the sports team.}

\item{year}{is a char string that corresponds to the playoff year.}

\item{stats_tb}{is a char string that corresponds to the statistics table such as #per_game, #totals, #per_36_minutes, and #advanced.}
}
\value{
a data frame of statistics for a specific NBA team.
}
\description{
Scrapes NBA team stats tables off basketball-reference.com.
}
\examples{
scrape_nba_team_stats("CHI", "1998","#totals")
}
